lambdas/rag-ingest/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
pypdf==3.17.0
numpy==1.26.2
